# refactor!: change decoder's listener behavior